Fixed backend image duplication++ (#110)
This PR includes several fixes to avoid file duplications and leftover
files in the backend. It fixes a bug where some image files that
should've been converted would cause an error. This PR also removes old
images when new ones are created, such as cropped images or
georeferenced images.

ImageHelper, image2png:
- Added a line to close image, fixes a bug where some files weren't
converted (like .tiff or .gif)
- Made renaming file account for different length file endings (.tiff =
4, .gif = 3)

ProjectHandler:
- For both non-referenced and georeferenced images, when updating file
path a check to see if an image already exists is made. If an image
exists, it is then deleted before the new one is uploaded
- Fixed a bug where some temporary files weren't deleted