Various frontend bugfixes and changes (#120)
This PR includes various bugfixes and changes in the frontend, such as:

**Editor (page, not component):**
- Removed image source check from this page (is moved to component)

**Editor (component):**
- Moved image source component here, is done when component mounts.
Redirects if image source is not valid/empty. Fixes a bug where a user
would be stuck in an Editor limbo state.
- Removed old "valid check" for georef. Is now instead told by
split-view that georef is valid when the georeffed image and its georef
coordinates are completed. Fixes a bug where the user would be placed
somewhere in Africa and potentially crashing the backend.

**projectAPI:**
- Added a variable reset. Fixes a bug where new projects would not be
added after a refresh, leaving the Editor in a limbo state.

**split-view:**
- Added a setHasBeenGeoreferenced-prop. Used for enabling Overlay and
Download buttons.

**EditorToolbar:**
- Added a warning popup when closing the browser/tab (The classic
"Changes you made may not be saved")