Extra Help Messages & Help Modal (#143)
This pull request adds some extra feedback for the user, and a help
modal, to attempt with making the application easier to use. In
addition, several recorded gifs explaining how to use the application
has been added.

Help messages:
As noted during user testing, some users were unaware of the points
needed for the georef to work. A counter has been added, such as "2 more
points needed to georeference".

Help modal:
To the left of the Feedback-button, there is now a Help button. This
opens up a modal containing several pages with information about how to
use the app, including:
- Introduction - What is georeferencing?
- Navigating around - Navbar **(has gif)**
- Navigating around - Split View **(has gif)**
- Navigating around - Overlay **(has gif)**
- Navigating around - Crop Mode **(has gif)**
- Adding Points (has gif)
- The Split View Toolbar
- The Overlay Toolbar
- The Coordinate Table **(has gif)**
- Help Modal
- Feedback Form
- Download File
- Exiting The Editor

---------

Co-authored-by: Meepception <35727117+Meepception@users.noreply.github.com>
Co-authored-by: Tom André Slåen Myre <tom.a.s.myre@gmail.com>